Fixed TB_05 back to its original state
It got some comments from TB_09 which I think
made it un-clear.
Also, the whole purpose of that example is to use Bloch's theorem.

Fixed wordings in TS_01 and TS_02 to clarify which steps are
actually needed in the exercises.

Added a fix in the installation script to make 0.14.3
ts-fft work.

Signed-off-by: Nick Papior <nickpapior@gmail.com>