Changes to TB_07; More significant, this time
Here I list a few suggestions that I think should be made at some point,
maybe for the next transiesta workshop?
- Perturbation terms (to H and self-energies) should be defined
somewhere
- I would add a couple of refs associated to the specific model utilized
to reproduce the effect of the magnetic field
- At some point it is mentioned that TBtrans does not allow complex
hamiltonians, but then in this tutorial we are building a complex
hamiltonian and providing it to TBtrans, which seems contradictory. I
understand that TBtrans cannot *build* complex Hamiltonians, but may
read them..? If so, that sentence should be modified for clarity.
- As we spoke about, I would not mention bond transmission maps if there
is no tutorial for them - I think for the general student that is
confusing. Again, for the next workshop it may be considered (maybe Tano
and I could do it... :D) to include one tutorial exclusively focused on
bond currents, and then mentioning bond currents here and there would
make full sense.
- I do not understand the "TIME:" and "TIME(HARD):" terms... What do you
mean with these?